Check that the schema is in agreement after execution of schema-realted commands
When using schema modification commands together with default load balancing policy, it's
possible to get schema disagreement between nodes of the cluster that will require a
separate set of steps to fix.

This PR checks if the executed command is related to schema change, and checks/waits for
schema agreement.